[Autotuner] LLM hybrid follow-ups
