+++ helion/autotuner/llm_seeded_lfbo.py
@@ -1,17 +1,29 @@
-"""Seed a second-stage autotuner with configs from an LLM search pass."""
+"""Run a two-stage hybrid autotuner that seeds a local search with an LLM pass.
+
+High-level flow:
+1. Run ``LLMGuidedSearch`` for ``llm_max_rounds`` rounds and keep its best
+   config. The hybrid defaults to 1 LLM round.
+2. Run a second-stage non-LLM search, ``LFBOTreeSearch`` by default.
+3. If the second stage supports best-available seeding, force
+   ``FROM_BEST_AVAILABLE`` and inject the LLM best config so stage 2 can refine
+   it instead of starting cold.
+4. Report per-stage timing and config-count metrics, plus aggregated hybrid
+   totals.
+
+Setting ``llm_max_rounds=0`` skips the LLM stage and runs only the second
+stage.
+"""
 
 from __future__ import annotations
 
-import inspect
 import math
 import os
 import time
 from typing import TYPE_CHECKING
-from typing import Any
 from typing import cast
 
 from .base_search import BaseSearch
-from .effort_profile import PATTERN_SEARCH_DEFAULTS
+from .base_search import PopulationBasedSearch
 from .effort_profile import QUICK_LLM_SEARCH_DEFAULTS
 from .llm.transport import DEFAULT_REQUEST_TIMEOUT_S
 from .llm_search import LLMGuidedSearch
@@ -19,6 +31,7 @@ from .llm_search import guided_search_kwargs_from_config
 from .pattern_search import InitialPopulationStrategy
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
     from collections.abc import Sequence
 
     from ..runtime.config import Config
@@ -32,11 +45,12 @@ _DISALLOWED_SECOND_STAGE_ALGORITHMS = {
     "LLMSeededSearch",
     "LLMSeededLFBOTreeSearch",
 }
-
-
-def _parse_env_bool(value: str) -> bool:
-    """Parse the small env-var bool dialect used by the hybrid overrides."""
-    return value.strip().lower() not in {"", "0", "false"}
+_AGGREGATED_METRIC_FIELDS = (
+    "num_configs_tested",
+    "num_compile_failures",
+    "num_accuracy_failures",
+    "num_generations",
+)
 
 
 def _resolve_second_stage_algorithm(name: str) -> type[BaseSearch]:
@@ -57,9 +71,12 @@ def _resolve_second_stage_algorithm(name: str) -> type[BaseSearch]:
     return search_cls
 
 
-def _supports_init_parameter(search_cls: type[BaseSearch], name: str) -> bool:
-    """Check whether a second-stage search accepts a particular kwarg."""
-    return name in inspect.signature(search_cls.__init__).parameters
+def _supports_best_available_handoff(search_cls: type[BaseSearch]) -> bool:
+    """Return whether the second stage supports FROM_BEST_AVAILABLE seeding."""
+    from .differential_evolution import DifferentialEvolutionSearch
+    from .pattern_search import PatternSearch
+
+    return issubclass(search_cls, (PatternSearch, DifferentialEvolutionSearch))
 
 
 class LLMSeededSearch(BaseSearch):
@@ -70,7 +87,7 @@ class LLMSeededSearch(BaseSearch):
     1. Run ``LLMGuidedSearch`` for ``llm_max_rounds`` rounds and capture its best
        config in memory.
     2. Run the configured second-stage search algorithm. If the algorithm
-       supports ``initial_population_strategy``, it is switched to
+       supports best-available seeding, it is switched to
        ``FROM_BEST_AVAILABLE`` so it can start from the LLM seed config.
 
     Setting ``llm_max_rounds=0`` disables the seed stage and runs only the
@@ -94,6 +111,7 @@ class LLMSeededSearch(BaseSearch):
         llm_configs_per_round: int = QUICK_LLM_SEARCH_DEFAULTS.configs_per_round,
         llm_max_rounds: int = QUICK_LLM_SEARCH_DEFAULTS.max_rounds,
         llm_initial_random_configs: int = QUICK_LLM_SEARCH_DEFAULTS.initial_random_configs,
+        llm_compile_timeout_s: int | None = QUICK_LLM_SEARCH_DEFAULTS.compile_timeout_s,
         llm_api_base: str | None = None,
         llm_api_key: str | None = None,
         llm_request_timeout_s: float = DEFAULT_REQUEST_TIMEOUT_S,
@@ -104,7 +122,12 @@ class LLMSeededSearch(BaseSearch):
         self.second_stage_algorithm = (
             second_stage_algorithm or type(self).default_second_stage_algorithm
         )
-        _resolve_second_stage_algorithm(self.second_stage_algorithm)
+        self._second_stage_search_cls = _resolve_second_stage_algorithm(
+            self.second_stage_algorithm
+        )
+        self._second_stage_supports_best_available_handoff = (
+            _supports_best_available_handoff(self._second_stage_search_cls)
+        )
         self.second_stage_kwargs = dict(second_stage_kwargs or {})
         self.best_available_pad_random = best_available_pad_random
 
@@ -113,6 +136,7 @@ class LLMSeededSearch(BaseSearch):
         self.llm_configs_per_round = llm_configs_per_round
         self.llm_max_rounds = llm_max_rounds
         self.llm_initial_random_configs = llm_initial_random_configs
+        self.llm_compile_timeout_s = llm_compile_timeout_s
         self.llm_api_base = llm_api_base
         self.llm_api_key = llm_api_key
         self.llm_request_timeout_s = llm_request_timeout_s
@@ -158,10 +182,6 @@ class LLMSeededSearch(BaseSearch):
 
         if (value := os.environ.get("HELION_HYBRID_LLM_MAX_ROUNDS")) is not None:
             kwargs["llm_max_rounds"] = int(value)
-        if (
-            value := os.environ.get("HELION_HYBRID_BEST_AVAILABLE_PAD_RANDOM")
-        ) is not None:
-            kwargs["best_available_pad_random"] = _parse_env_bool(value)
         return kwargs
 
     def _make_llm_search(self) -> LLMGuidedSearch:
@@ -175,37 +195,39 @@ class LLMSeededSearch(BaseSearch):
             configs_per_round=self.llm_configs_per_round,
             max_rounds=self.llm_max_rounds,
             initial_random_configs=self.llm_initial_random_configs,
+            compile_timeout_s=self.llm_compile_timeout_s,
             api_base=self.llm_api_base,
             api_key=self.llm_api_key,
             request_timeout_s=self.llm_request_timeout_s,
         )
 
+    def _second_stage_search_kwargs(self, *, seeded: bool) -> dict[str, object]:
+        """Build the stage-2 kwargs, forcing best-available seeding when supported."""
+        kwargs = dict(self.second_stage_kwargs)
+        if not seeded:
+            return kwargs
+
+        if not self._second_stage_supports_best_available_handoff:
+            self.log(
+                f"Second-stage algorithm {self.second_stage_algorithm} "
+                "does not support FROM_BEST_AVAILABLE initialization; "
+                "the LLM seed may not influence the next stage."
+            )
+            return kwargs
+
+        kwargs["initial_population_strategy"] = (
+            InitialPopulationStrategy.FROM_BEST_AVAILABLE
+        )
+        kwargs["best_available_pad_random"] = self.best_available_pad_random
+        return kwargs
+
     def _make_second_stage_search(self, *, seeded: bool) -> BaseSearch:
         """Construct stage 2 and enable best-available seeding when supported."""
-        second_stage_cls = _resolve_second_stage_algorithm(self.second_stage_algorithm)
-        kwargs = dict(self.second_stage_kwargs)
-
-        if seeded:
-            if _supports_init_parameter(
-                second_stage_cls, "initial_population_strategy"
-            ):
-                kwargs["initial_population_strategy"] = (
-                    InitialPopulationStrategy.FROM_BEST_AVAILABLE
-                )
-                if _supports_init_parameter(
-                    second_stage_cls, "best_available_pad_random"
-                ):
-                    kwargs["best_available_pad_random"] = self.best_available_pad_random
-            else:
-                self.log(
-                    f"Second-stage algorithm {self.second_stage_algorithm} "
-                    "does not support FROM_BEST_AVAILABLE initialization; "
-                    "the LLM seed may not influence the next stage."
-                )
-
-        return cast(
-            "BaseSearch",
-            cast("Any", second_stage_cls)(self.kernel, self.args, **kwargs),
+        factory = cast("Callable[..., BaseSearch]", self._second_stage_search_cls)
+        return factory(
+            self.kernel,
+            self.args,
+            **self._second_stage_search_kwargs(seeded=seeded),
         )
 
     def _inject_seed_into_second_stage(
@@ -214,10 +236,57 @@ class LLMSeededSearch(BaseSearch):
         llm_seed_config: Config,
     ) -> None:
         """Pass the best LLM config into searches that expose the seed hook."""
-        setter = getattr(second_stage_search, "set_best_available_seed_configs", None)
-        if setter is None:
+        if not self._second_stage_supports_best_available_handoff:
             return
-        setter([llm_seed_config])
+        seeded_search = cast("PopulationBasedSearch", second_stage_search)
+        seeded_search.set_best_available_seed_configs([llm_seed_config])
+
+    @staticmethod
+    def _finite_perf(search: BaseSearch | None) -> float | None:
+        """Return a search's best perf when finite, else None for reporting."""
+        if search is None or not math.isfinite(search.best_perf_so_far):
+            return None
+        return search.best_perf_so_far
+
+    def _run_llm_seed_stage(
+        self,
+    ) -> tuple[LLMGuidedSearch | None, Config | None, float]:
+        """Run the optional stage-1 LLM search and return its best config."""
+        if self.llm_max_rounds <= 0:
+            return None, None, 0.0
+
+        self.log(
+            "Hybrid stage 1/2: "
+            f"LLMGuidedSearch for {self.llm_max_rounds} round(s) "
+            f"with {self.llm_configs_per_round} configs/round"
+        )
+        llm_search = self._make_llm_search()
+        llm_start = time.perf_counter()
+        llm_seed_config = llm_search.autotune(skip_cache=True)
+        llm_wall_time = time.perf_counter() - llm_start
+        return llm_search, llm_seed_config, llm_wall_time
+
+    def _run_second_stage(
+        self,
+        llm_seed_config: Config | None,
+    ) -> tuple[BaseSearch, Config, float]:
+        """Run stage 2, optionally seeded from the stage-1 best config."""
+        seeded = llm_seed_config is not None
+        self.log(
+            "Hybrid stage 2/2: "
+            + (
+                f"running {self.second_stage_algorithm} from best available seed"
+                if seeded
+                else f"running {self.second_stage_algorithm} without LLM seed"
+            )
+        )
+        second_stage_search = self._make_second_stage_search(seeded=seeded)
+        if llm_seed_config is not None:
+            self._inject_seed_into_second_stage(second_stage_search, llm_seed_config)
+        second_stage_start = time.perf_counter()
+        best_config = second_stage_search.autotune()
+        second_stage_wall_time = time.perf_counter() - second_stage_start
+        return second_stage_search, best_config, second_stage_wall_time
 
     def _finalize_stage_metrics(
         self,
@@ -229,18 +298,13 @@ class LLMSeededSearch(BaseSearch):
     ) -> None:
         """Merge per-stage timing and autotune metrics into the hybrid summary."""
 
-        def _finite_perf(search: BaseSearch | None) -> float | None:
-            if search is None or not math.isfinite(search.best_perf_so_far):
-                return None
-            return search.best_perf_so_far
-
         llm_metrics = llm_search._autotune_metrics if llm_search else None
         second_stage_metrics = second_stage_search._autotune_metrics
         second_stage_tested = second_stage_metrics.num_configs_tested
 
         self.hybrid_stage_breakdown = {
             "used_llm_seed": llm_search is not None,
-            "llm_seed_perf_ms": _finite_perf(llm_search),
+            "llm_seed_perf_ms": self._finite_perf(llm_search),
             "llm_seed_time_s": llm_wall_time,
             "llm_seed_configs_tested": (
                 llm_metrics.num_configs_tested if llm_metrics else 0
@@ -249,26 +313,13 @@ class LLMSeededSearch(BaseSearch):
                 dict(llm_seed_config) if llm_seed_config is not None else None
             ),
             "second_stage_algorithm": self.second_stage_algorithm,
-            "second_stage_perf_ms": _finite_perf(second_stage_search),
+            "second_stage_perf_ms": self._finite_perf(second_stage_search),
             "second_stage_time_s": second_stage_wall_time,
             "second_stage_configs_tested": second_stage_tested,
         }
-        if self.second_stage_algorithm == "LFBOTreeSearch":
-            self.hybrid_stage_breakdown.update(
-                {
-                    "lfbo_stage_perf_ms": _finite_perf(second_stage_search),
-                    "lfbo_stage_time_s": second_stage_wall_time,
-                    "lfbo_stage_configs_tested": second_stage_tested,
-                }
-            )
 
         # Aggregate metrics from both stages
-        for field in (
-            "num_configs_tested",
-            "num_compile_failures",
-            "num_accuracy_failures",
-            "num_generations",
-        ):
+        for field in _AGGREGATED_METRIC_FIELDS:
             setattr(
                 self._autotune_metrics,
                 field,
@@ -293,38 +344,12 @@ class LLMSeededSearch(BaseSearch):
             f"best_available_pad_random={self.best_available_pad_random}"
         )
 
-        # Stage 1: LLM seed search
-        llm_search: LLMGuidedSearch | None = None
-        llm_seed_config: Config | None = None
-        llm_wall_time = 0.0
-
-        if self.llm_max_rounds > 0:
-            self.log(
-                "Hybrid stage 1/2: "
-                f"LLMGuidedSearch for {self.llm_max_rounds} round(s) "
-                f"with {self.llm_configs_per_round} configs/round"
-            )
-            llm_search = self._make_llm_search()
-            llm_start = time.perf_counter()
-            llm_seed_config = llm_search.autotune(skip_cache=True)
-            llm_wall_time = time.perf_counter() - llm_start
-
-        # Stage 2: second-stage search (optionally seeded)
-        seeded = llm_seed_config is not None
-        self.log(
-            "Hybrid stage 2/2: "
-            + (
-                f"running {self.second_stage_algorithm} from best available seed"
-                if seeded
-                else f"running {self.second_stage_algorithm} without LLM seed"
-            )
+        # Stage 1: run the LLM seed search when enabled and keep its best config.
+        llm_search, llm_seed_config, llm_wall_time = self._run_llm_seed_stage()
+        # Stage 2: run the configured follow-up search, seeded when stage 1 found a config.
+        second_stage_search, best_config, second_stage_wall_time = (
+            self._run_second_stage(llm_seed_config)
         )
-        second_stage_search = self._make_second_stage_search(seeded=seeded)
-        if llm_seed_config is not None:
-            self._inject_seed_into_second_stage(second_stage_search, llm_seed_config)
-        second_stage_start = time.perf_counter()
-        best_config = second_stage_search.autotune()
-        second_stage_wall_time = time.perf_counter() - second_stage_start
 
         self._finalize_stage_metrics(
             llm_search,
@@ -337,7 +362,10 @@ class LLMSeededSearch(BaseSearch):
 
 
 class LLMSeededLFBOTreeSearch(LLMSeededSearch):
-    """Convenience wrapper for the common LLM-seeded LFBO tree search pipeline."""
+    """Convenience wrapper for the common LLM-seeded LFBO tree search pipeline.
+
+    LFBO-specific stage-2 settings should be passed through ``second_stage_kwargs``.
+    """
 
     allow_second_stage_env_override = False
 
@@ -356,63 +384,29 @@ class LLMSeededLFBOTreeSearch(LLMSeededSearch):
         args: Sequence[object],
         *,
         second_stage_kwargs: dict[str, object] | None = None,
-        num_neighbors: int = 200,
-        frac_selected: float = 0.10,
-        radius: int = 2,
-        initial_population: int = PATTERN_SEARCH_DEFAULTS.initial_population,
-        copies: int = PATTERN_SEARCH_DEFAULTS.copies,
-        max_generations: int = PATTERN_SEARCH_DEFAULTS.max_generations,
-        min_improvement_delta: float = 0.001,
-        quantile: float = 0.1,
-        patience: int = 1,
-        similarity_penalty: float = 1.0,
-        initial_population_strategy: InitialPopulationStrategy | None = None,
         best_available_pad_random: bool = False,
-        finishing_rounds: int = 0,
-        compile_timeout_lower_bound: float = PATTERN_SEARCH_DEFAULTS.compile_timeout_lower_bound,
-        compile_timeout_quantile: float = PATTERN_SEARCH_DEFAULTS.compile_timeout_quantile,
         llm_provider: str | None = None,
         llm_model: str = QUICK_LLM_SEARCH_DEFAULTS.model,
         llm_configs_per_round: int = QUICK_LLM_SEARCH_DEFAULTS.configs_per_round,
         llm_max_rounds: int = QUICK_LLM_SEARCH_DEFAULTS.max_rounds,
         llm_initial_random_configs: int = QUICK_LLM_SEARCH_DEFAULTS.initial_random_configs,
+        llm_compile_timeout_s: int | None = QUICK_LLM_SEARCH_DEFAULTS.compile_timeout_s,
         llm_api_base: str | None = None,
         llm_api_key: str | None = None,
         llm_request_timeout_s: float = DEFAULT_REQUEST_TIMEOUT_S,
     ) -> None:
-        # Build LFBO second-stage kwargs from individual params or passthrough
-        computed_second_stage_kwargs: dict[str, object]
-        if second_stage_kwargs is not None:
-            computed_second_stage_kwargs = dict(second_stage_kwargs)
-        else:
-            computed_second_stage_kwargs = {
-                "num_neighbors": num_neighbors,
-                "frac_selected": frac_selected,
-                "radius": radius,
-                "initial_population": initial_population,
-                "copies": copies,
-                "max_generations": max_generations,
-                "min_improvement_delta": min_improvement_delta,
-                "quantile": quantile,
-                "patience": patience,
-                "similarity_penalty": similarity_penalty,
-                "initial_population_strategy": initial_population_strategy,
-                "finishing_rounds": finishing_rounds,
-                "compile_timeout_lower_bound": compile_timeout_lower_bound,
-                "compile_timeout_quantile": compile_timeout_quantile,
-            }
-
         super().__init__(
             kernel,
             args,
             second_stage_algorithm="LFBOTreeSearch",
-            second_stage_kwargs=computed_second_stage_kwargs,
+            second_stage_kwargs=second_stage_kwargs,
             best_available_pad_random=best_available_pad_random,
             llm_provider=llm_provider,
             llm_model=llm_model,
             llm_configs_per_round=llm_configs_per_round,
             llm_max_rounds=llm_max_rounds,
             llm_initial_random_configs=llm_initial_random_configs,
+            llm_compile_timeout_s=llm_compile_timeout_s,
             llm_api_base=llm_api_base,
             llm_api_key=llm_api_key,
             llm_request_timeout_s=llm_request_timeout_s,
